core/settings: extract FlowSettings into @holons/core
Move FlowSettings, federation link helpers, and the underlying
loadSettings/saveSettings primitives into @holons/core/settings so
the Harvest web app and the Telegram bot share one persistence layer
for holon settings.

- packages/core/src/settings/{index,flow-settings,federation-links,equation}.ts
- apps/web/src/lib/holons/FlowSettings.ts is now a thin re-export
- packages/telegram-ui/src/Settings.js routes db.get/db.put('settings')
  through coreLoadSettings/coreSaveSettings while keeping all Telegraf
  command wiring intact
- packages/core/package.json exposes ./settings (and other subpaths)
  via dist with explicit types/import conditions so JS consumers
  (the bot) can import the built output and TS consumers still get
  proper type resolution
- equation.ts is intentionally a stub; loadEquation/subscribeToEquation
  live next to ScoreEquation in ContributionScoring.ts and are owned
  by the core/scoring unit
